refactor: useModal 커스텀 훅을 hooks 폴더 하위로 이동
